+++ packages/messenger-cli/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0]
+
 ### Added
 
 - **BREAKING:** Add support for formatting the generated method action type files with Prettier or Oxfmt ([#8486](https://github.com/MetaMask/core/pull/8486))
@@ -23,5 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - CLI tool for generating TypeScript action type files for controllers and services that define `MESSENGER_EXPOSED_METHODS`.
   - Available as a CLI binary (`messenger-action-types`).
 
-[Unreleased]: https://github.com/MetaMask/core/compare/@metamask/messenger-cli@0.1.0...HEAD
+[Unreleased]: https://github.com/MetaMask/core/compare/@metamask/messenger-cli@0.2.0...HEAD
+[0.2.0]: https://github.com/MetaMask/core/compare/@metamask/messenger-cli@0.1.0...@metamask/messenger-cli@0.2.0
 [0.1.0]: https://github.com/MetaMask/core/releases/tag/@metamask/messenger-cli@0.1.0